Added tests for standard worker functions & separated the commons
